fix duration variables from long to double

+++ src/main/java/sg/nus/iss/leavesystem/ca/model/Staff.java
@@ -233,7 +233,7 @@ public class Staff {
 	}
 
 	public void deductLeave(LeaveApplication leaveApplication){
-		long duration = Long.parseLong(leaveApplication.getDuration());
+		double duration = Double.parseDouble(leaveApplication.getDuration());
 		if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "annual"){
 			annualLeaveBalance-= duration;
 		}else if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "medical"){
@@ -244,7 +244,7 @@ public class Staff {
 	}
 
 	public void reinstateLeaveBalance(LeaveApplication leaveApplication){
-		long duration = Long.parseLong(leaveApplication.getDuration());
+		double duration = Double.parseDouble(leaveApplication.getDuration());
 		if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "annual"){
 			annualLeaveBalance+= duration;
 		}else if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "medical"){
@@ -255,7 +255,7 @@ public class Staff {
 	}
 
 	public void reinstatePreviousLeaveBalance(LeaveApplication leaveApplication,String durationStr){
-		long duration = Long.parseLong(durationStr);
+		double duration = Double.parseDouble(durationStr);
 		if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "annual"){
 			annualLeaveBalance+= duration;
 		}else if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "medical"){
@@ -267,7 +267,7 @@ public class Staff {
 
 	public Boolean isLeaveBalanceEnough(LeaveApplication leaveApplication)
 	{
-		long duration = Long.parseLong(leaveApplication.getDuration());
+		double duration = Double.parseDouble(leaveApplication.getDuration());
 		if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "annual"){
 			if(duration > annualLeaveBalance) return false;
 		}else if(leaveApplication.getTypeOfLeave().getLeaveTypeName() == "medical"){
